Refactor for window events
